Fix iOS code path (issues: #324, #294) 	Now plugin works on ios devices in all the cases (Foreground, Background, and even after Terminating the app).
Add LOCATION_ALWAYS_AND_WHEN_IN_USE_USAGE_DESCRIPTION in plugin.xml to support background location tracking in iOS-11

+++ ios/CDVBackgroundGeolocation/SQLiteLocationDAO.m
@@ -261,7 +261,7 @@
             return;
         } else if (rowCount > maxRows) {
             sql = [NSString stringWithFormat:@"DELETE FROM %1$@ WHERE %2$@ IN (SELECT %2$@ FROM %1$@ ORDER BY %3$@ LIMIT %4$ld);VACUUM;",
-                   @LC_TABLE_NAME, @LC_COLUMN_NAME_ID, @LC_COLUMN_NAME_TIME, (rowCount - maxRows)];
+                  @LC_TABLE_NAME, @LC_COLUMN_NAME_ID, @LC_COLUMN_NAME_TIME, (long)(rowCount - maxRows)];
             if (![database executeStatements:sql]) {
                 NSLog(@"%@ failed code: %d: message: %@", sql, [database lastErrorCode], [database lastErrorMessage]);
             }
